adds delete button functionality

+++ back-end/controllers/posts.js
@@ -37,6 +37,9 @@ export const likePost = async (req, res) => {
 export const deletePost = async (req, res) => {
     const { id } = req.params;
 
+    if (!mongoose.Types.ObjectId.isValid(id)) return res.status(404).send("No post with that id")
+
     await PostMessage.findByIdAndRemove(id)
+
     res.json({ message: "Post was deleted" })
 }
